Add files via upload
Pick2Light is an is client server concept used in warehouse management. This helps the industries to
increase the efficiency, instead of paper base approach of sorting, ordering and picking components.
The idea is to direct the floor operator to the glowing light and pick the displayed amount of
component from the bin.